+++ imgaug/__init__.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import
 from imgaug.imgaug import *
+import imgaug.augmentables as augmentables
+from imgaug.augmentables import *
 import imgaug.augmenters as augmenters
 import imgaug.parameters as parameters
 import imgaug.dtypes as dtypes