+++ imgaug/augmenters/segmentation.py
@@ -24,7 +24,7 @@ from skimage import segmentation, measure
 import six.moves as sm
 
 from . import meta
-from .. import imgaug as ia
+import imgaug as ia
 from .. import parameters as iap
 from .. import dtypes as iadt