+++ imgaug/augmenters/size.py
@@ -32,7 +32,7 @@ import numpy as np
 import six.moves as sm
 
 from . import meta
-from .. import imgaug as ia
+import imgaug as ia
 from .. import parameters as iap