+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 * Changed check in `HeatmapsOnImage` that validates whether the input array is within the desired value range `[min_value, max_value]` 
   from a hard exception to a soft warning (with clipping). Also improved the error message a bit.
 * Fixed assert in `SegmentationMapOnImage` falsely checking if max class index is `<= nb_classes` instead of `< nb_classes`. 
+* Moved `Keypoint`, `KeypointsOnImage` and `imgaug.imgaug.compute_geometric_median` to `augmentables/kps.py`.
 
 
 # 0.2.8